feat(cardano-chain-follower): Split reader and follower and reader now supports reading single block and block ranges from Mithril snapshots
